pyproject.toml
# PROJECT SETUP

[tool.poetry]
name = "lucky_forums"
version = "0.1.1"
description = "A forum-like service."
authors = ["Islam Yermek 35352@iitu.edu.kz"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "3.11.14"

# CORE FRAMEWORK

django = "^4.2"
djangorestframework = "^3.14"
psycopg2-binary = "^2.9.9"
gunicorn = "^21.2.0"
uvicorn = "^0.23.2"

# ASYNCHRONOUS TASKS & SCHEDULING

celery = "^5.3.4"
redis = "^5.0.1"
django-celery-beat = "^2.5.0"

# AUTHENTICATION & SECURITY

djangorestframework-simplejwt = "^5.3.0"
django-cors-headers = "^4.2.0"

# FILE HANDLING & STORAGE

boto3 = "^1.28.64"

django-storages = "^1.14"

# PERMISSIONS & RATE LIMITING

django-guardian = "^2.4.0"
django-ratelimit = "^4.1.0"

# DATABASE & MODELS

django-safedelete = "^1.3.2"


# UTILITIES

python-decouple = "^3.8"
shortuuid = "^1.0.11"
django-filter = "^23.3"
drf-nested-routers = "^0.95.0"
pillow = "^12.0.0"
markdown = "^3.7"
bleach = "^6.2.0"

[tool.poetry.group.dev.dependencies]

# CODE QUALITY & FORMATTING

black = "^23.9.1"
isort = "^5.12.0"
flake8 = "^6.1.0"

# TESTING

pytest = "^7.4.2"
pytest-django = "^4.7.0"
pytest-cov = "^4.1.0"
model-bakery = "^1.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py311"]
include = "\\.pyi?$"
extend-exclude = """
(
  ^/.venv
  | ^/build
  | ^/dist
  | ^/node_modules
  | ^/.mypy_cache
  | ^/.pytest_cache
  | ^/.ruff_cache
  | migrations
)
"""

[tool.isort]
profile = "black"
